requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
Babel==2.6.0
cachetools==3.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
click==6.7
cryptography==2.6.1
defusedxml==0.6.0
dj-database-url==0.4.2
Django==2.1.7
django-address==0.2.0
django-allauth==0.22.0
django-autocomplete-light==3.3.2
django-autoslug-iplweb==1.9.4.dev0
django-cities-light==3.5.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-phonenumber-field==2.2.0
django-search-views==0.3.7
futures==3.1.1
google-assistant-grpc==0.2.1
google-assistant-sdk==0.5.1
google-auth==1.6.3
google-auth-oauthlib==0.3.0
googleapis-common-protos==1.5.10
grpcio==1.20.1
gunicorn==19.7.1
idna==2.8
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
oauthlib==3.0.1
pathlib2==2.3.3
phonenumbers==8.10.6
Pillow==6.0.0
progressbar2==3.39.3
protobuf==3.7.1
psycopg2-binary==2.8.2
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
pylint==2.3.1
pyOpenSSL==19.0.0
python-utils==2.3.0
python3-openid==3.1.0
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
sounddevice==0.3.13
sqlparse==0.3.0
tenacity==4.12.0
typed-ast==1.3.5
Unidecode==1.0.23
urllib3==1.24.3
virtualenv==15.1.0
whitenoise==3.3.1
wrapt==1.11.1
